requirements.txt
Flask==3.0.3
colorama==0.4.6
python-nmap==0.7.1
scapy==2.7.0
